# Go: reduce copying of orders