Performance improvement about 11%.

This version:

% go build && for i in $(seq 5) ; gzip -dc ../../orders-10M.txt.gz | time ./Exchange >/dev/null
./Exchange > /dev/null  9.92s user 0.63s system 148% cpu 7.084 total
./Exchange > /dev/null  10.46s user 0.62s system 153% cpu 7.224 total
./Exchange > /dev/null  9.72s user 0.65s system 146% cpu 7.076 total
./Exchange > /dev/null  9.86s user 0.64s system 151% cpu 6.948 total
./Exchange > /dev/null  9.86s user 0.65s system 151% cpu 6.958 total

Previous version:

% go build && for i in $(seq 5) ; gzip -dc ../../orders-10M.txt.gz | time ./Exchange >/dev/null
./Exchange > /dev/null  13.42s user 0.67s system 176% cpu 7.966 total
./Exchange > /dev/null  14.31s user 0.71s system 188% cpu 7.985 total
./Exchange > /dev/null  14.79s user 0.74s system 192% cpu 8.087 total
./Exchange > /dev/null  14.39s user 0.70s system 190% cpu 7.905 total
./Exchange > /dev/null  13.86s user 0.69s system 183% cpu 7.946 total